added a config file loader plus 12 dependencies
Added a config file loader. The config file for tdk can be any of these extensions:

* `.ini` or `.config` - ini@1.3.3
* `.yml` or `.yaml` - js-yaml@3.2.7
* `.json` - strip-json-comments@1.0.2 & native JSON
* `.json5` - json5@0.4.0
* `.hjson` - hjson@1.6.1
* `.toml` - toml@2.1.1
* `.cson` - cson@2.0.0
* `.properties` - properties@1.2.1
* `.coffee`, `.litcoffee` or `.coffee.md` - coffee-script@1.9.1
* `.iced`, `.liticed` or `.iced.md` - iced-coffee-script@1.8.0-c
* `.oli` - oli@0.1.0
* `.xml` - xml2js@0.4.5
* `.js` or `.*` - attempts to use `require`

If any of the loaders or parsers return a function, the function will be called with the tdk app as its scope (`this` object) and the parsed argv and tdk app as the arguments.
After that (or if it wasn't a function) the object must be a JavaScript Hash Object (`{ ..., /* key: value */  ... }`)